Extract the MRU management from MruCrlCache
