+++ src/libraries/System.Security.Cryptography/src/System/Security/Cryptography/X509Certificates/OpenSslCrlCache.cs
@@ -483,105 +483,33 @@ namespace System.Security.Cryptography.X509Certificates
         // The MRU CRL cache always does a DangerousAddReference before returning the value,
         // so that neither cooperative GC pruning nor a cache-value refresh trigger ReleaseHandle
         // on a CRL entry in use.
-        private sealed class MruCrlCache
+        private sealed class MruCrlCache : X509MruCache<CachedCrlEntry>
         {
-            // Each CRL is only a SafeHandle to the GC, but represents a non-trivial amount of
-            // native memory, so keep the cache small.
-            private const int MaxItems = 30;
-
-            private readonly Lock _lock = new();
-
-            private int _count = -1;
-            private Node? _head;
-            private Node? _expire;
-
             internal CachedCrlEntry AddOrUpdateAndUpRef(string key, CachedCrlEntry value)
             {
-                Debug.Assert(key is not null);
-                Debug.Assert(value is not null);
-                Debug.Assert(value.CrlHandle is not null && !value.CrlHandle.IsInvalid);
-                // Don't assert/enforce anything about expiration, because a) clock-skew, or b)
-                // the caller might have a verification time that's in the past.
-
-                int hashCode = key.GetHashCode();
-                CachedCrlEntry ret = value;
-                string? fullMemberKey = null;
-                SafeX509CrlHandle? toDispose = null;
+                CachedCrlEntry ret;
+                Node? evicted;
+                CachedCrlEntry? replaced;
+                int hashCode = GetHashCode(key);
 
                 lock (_lock)
                 {
-                    // The first time we add something, create the object to monitor for GC events.
-                    if (_count < 0)
-                    {
-                        new GCWatcher(this);
-                        _count = 0;
-                    }
+                    ret = AddOrUpdate(hashCode, key, value, out evicted, out replaced);
 
                     bool ignore = false;
-
-                    if (TryGetNode(hashCode, key, out Node? current))
-                    {
-                        Debug.Assert(current is not null);
-
-                        if (current.Value.Expiration >= value.Expiration)
-                        {
-                            toDispose = value.CrlHandle;
-                            ret = current.Value;
-                        }
-                        else
-                        {
-                            toDispose = current.Value.CrlHandle;
-                            current.Value = value;
-                        }
-                    }
-                    else
-                    {
-                        Node node = new Node(hashCode, key, value);
-                        node.Next = _head;
-
-                        if (_count < MaxItems)
-                        {
-                            _count++;
-                        }
-                        else
-                        {
-                            // Because MaxItems is small, it's better to just iterate from head
-                            // instead of using a doubly-linked list.
-
-                            Node? previous = null;
-                            Node? cur = _head;
-                            Node? next = cur?.Next;
-
-                            while (next is not null)
-                            {
-                                previous = cur;
-                                cur = next;
-                                next = cur.Next;
-                            }
-
-                            Debug.Assert(previous is not null);
-                            Debug.Assert(cur is not null);
-
-                            previous.Next = null;
-                            toDispose = cur.Value.CrlHandle;
-                            fullMemberKey = cur.Key;
-                            if (cur == _expire)
-                            {
-                                _expire = null;
-                            }
-                        }
-
-                        _head = node;
-                    }
-
                     ret.CrlHandle.DangerousAddRef(ref ignore);
                 }
 
-                toDispose?.Dispose();
+                replaced?.CrlHandle.Dispose();
 
-                if (fullMemberKey is not null && OpenSslX509ChainEventSource.Log.IsEnabled())
+                if (evicted is not null)
                 {
-                    OpenSslX509ChainEventSource.Log.CrlCacheInMemoryFull(fullMemberKey);
+                    evicted.Value.CrlHandle.Dispose();
+
+                    if (OpenSslX509ChainEventSource.Log.IsEnabled())
+                    {
+                        OpenSslX509ChainEventSource.Log.CrlCacheInMemoryFull(evicted.Key);
+                    }
                 }
 
                 return ret;
@@ -589,7 +517,7 @@ namespace System.Security.Cryptography.X509Certificates
 
             internal bool TryGetValueAndUpRef(string key, [NotNullWhen(true)] out CachedCrlEntry? value)
             {
-                int hashCode = key.GetHashCode();
+                int hashCode = GetHashCode(key);
 
                 lock (_lock)
                 {
@@ -606,159 +534,20 @@ namespace System.Security.Cryptography.X509Certificates
                 return false;
             }
 
-            private bool TryGetNode(int hashCode, string key, [NotNullWhen(true)] out Node? value)
+            private protected override bool OnConflictTakeNew(Node current, CachedCrlEntry newValue) => newValue.Expiration > current.Value.Expiration;
+
+            private protected override void Pruned(Node? prunedNode, int countStart, int countEnd)
             {
-                Debug.Assert(_lock.IsHeldByCurrentThread);
-
-                Node? previous = null;
-                Node? current = _head;
-
-                while (current is not null)
+                // `prunedNode` and beyond are now unlinked from the list, so we can dispose its values without holding the lock.
+                while (prunedNode is not null)
                 {
-                    if (current.MatchesKey(hashCode, key))
-                    {
-                        // If we find the expire node, move expiration to after it, so that promoting it to
-                        // most recent doesn't prune the whole list.
-                        //
-                        // This might, of course, make _expire null.
-                        if (current == _expire)
-                        {
-                            _expire = current.Next;
-                        }
-
-                        // Move the found node to the head of the list, maintaining MRU ordering.
-                        if (previous != null)
-                        {
-                            previous.Next = current.Next;
-                            current.Next = _head;
-                            _head = current;
-                        }
-
-                        value = current;
-                        return true;
-                    }
-
-                    previous = current;
-                    current = current.Next;
-                }
-
-                value = null;
-                return false;
-            }
-
-            private void PruneForGC()
-            {
-                // The general flow:
-                // * The current head is where we expire next time.
-                // * Under the lock: If there is an expire node, determine the new count by walking to it,
-                //   and unlink it from the previous node.
-                // * After the lock: Dispose all the values from the prune node onward.
-
-                Node? prune;
-                int countStart;
-                int countEnd;
-
-                lock (_lock)
-                {
-                    prune = _expire;
-                    _expire = _head;
-                    countStart = _count;
-
-                    if (prune is null)
-                    {
-                        return;
-                    }
-
-                    if (prune == _head)
-                    {
-                        _count = 0;
-                        _head = null;
-                        _expire = null;
-                    }
-                    else
-                    {
-                        Debug.Assert(_head is not null);
-                        int count = 1;
-                        Node current = _head;
-
-                        while (current.Next != prune && current.Next is not null)
-                        {
-                            count++;
-                            current = current.Next;
-                        }
-
-                        Debug.Assert(current.Next == prune, "The prune node should be in the list");
-                        current.Next = null;
-                        _count = count;
-                    }
-
-                    countEnd = _count;
-                }
-
-                // `prune` and beyond are now unlinked from the list, so we can dispose its values without holding the lock.
-                while (prune is not null)
-                {
-                    prune.Value.CrlHandle.Dispose();
-                    prune = prune.Next;
+                    prunedNode.Value.CrlHandle.Dispose();
+                    prunedNode = prunedNode.Next;
                 }
 
                 if (OpenSslX509ChainEventSource.Log.IsEnabled())
                 {
                     OpenSslX509ChainEventSource.Log.CrlCacheInMemoryPruned(countStart - countEnd, countEnd);
-                }
-            }
-
-            private sealed class Node
-            {
-                private readonly int _keyHashCode;
-
-                internal string Key { get; }
-                internal CachedCrlEntry Value { get; set; }
-                internal Node? Next { get; set; }
-
-                internal Node(int hashCode, string key, CachedCrlEntry value)
-                {
-                    Debug.Assert(key.GetHashCode() == hashCode);
-
-                    Key = key;
-                    _keyHashCode = hashCode;
-                    Value = value;
-                }
-
-                internal bool MatchesKey(int hashCode, string key)
-                {
-                    return _keyHashCode == hashCode && Key.Equals(key, StringComparison.Ordinal);
-                }
-            }
-
-            private sealed class GCWatcher
-            {
-                private readonly MruCrlCache _owner;
-
-                internal GCWatcher(MruCrlCache owner)
-                {
-                    _owner = owner;
-                }
-
-                ~GCWatcher()
-                {
-                    GC.ReRegisterForFinalize(this);
-
-                    if (GC.GetGeneration(this) == GC.MaxGeneration)
-                    {
-                        try
-                        {
-                            _owner.PruneForGC();
-                        }
-                        catch
-                        {
-                            // Eat any exception so we don't terminate the finalizer thread.
-#if DEBUG
-                            // Except in DEBUG, as we really shouldn't be hitting any exceptions here.
-                            throw;
-#endif
-                        }
-                    }
                 }
             }
         }
